multi platform beta/requirements.txt
numpy
pygame
PyOpenGL
opencv-python
Pillow
pyautogui